+++ spine-ios/Example/Spine iOS Example/AnimationStateEvents.swift
@@ -58,4 +58,3 @@ struct AnimationStateEvents: View {
 #Preview {
     AnimationStateEvents()
 }
-