+++ spine-ios/Example/Spine iOS Example/spine-ios/SpineRenderer.swift
@@ -19,7 +19,7 @@ protocol SpineRendererDelegate: AnyObject {
     func spineRendererWillDraw(_ spineRenderer: SpineRenderer)
     func spineRendererDidDraw(_ spineRenderer: SpineRenderer)
     
-    func spineRendererDidUpdate(_ spineRenderer: SpineRenderer, scaleX: CGFloat, scaleY: CGFloat, offsetX: CGFloat, offsetY: CGFloat)
+    func spineRendererDidUpdate(_ spineRenderer: SpineRenderer, scaleX: CGFloat, scaleY: CGFloat, offsetX: CGFloat, offsetY: CGFloat, size: CGSize)
 }
 
 protocol SpineRendererDataSource: AnyObject {
@@ -121,8 +121,8 @@ final class SpineRenderer: NSObject, MTKViewDelegate {
     }
     
     private func setTransform(bounds: CGRect, mode: Spine.ContentMode, alignment: Spine.Alignment) {
-        let x = -bounds.minX - bounds.width / 2.0// - (alignment.x * bounds.width / 2.0)
-        let y = -bounds.minY - bounds.height / 2.0// - (alignment.y * bounds.height / 2.0)
+        let x = -bounds.minX - bounds.width / 2.0
+        let y = -bounds.minY - bounds.height / 2.0
         
         var scaleX: CGFloat = 1.0
         var scaleY: CGFloat = 1.0
@@ -136,11 +136,8 @@ final class SpineRenderer: NSObject, MTKViewDelegate {
             scaleY = scaleX
         }
         
-        let offsetX = abs(sizeInPoints.width - bounds.width) / 2 * alignment.x
-        let offsetY = abs(sizeInPoints.height - bounds.height) / 2 * alignment.y
-        
-//        let offsetX = sizeInPoints.width / 2.0 + (alignment.x * sizeInPoints.width / 2.0)
-//        let offsetY = sizeInPoints.height / 2.0 + (alignment.y * sizeInPoints.height / 2.0)
+        let offsetX = abs(sizeInPoints.width - bounds.width * scaleX) / 2 * alignment.x
+        let offsetY = abs(sizeInPoints.height - bounds.height * scaleY) / 2 * alignment.y
         
         transform = AAPLTransform(
             translation: vector_float2(Float(x), Float(y)),
@@ -152,8 +149,9 @@ final class SpineRenderer: NSObject, MTKViewDelegate {
             self,
             scaleX: scaleX,
             scaleY: scaleY,
-            offsetX: x / scaleX + bounds.width / 2.0 + offsetX,
-            offsetY: y / scaleY + bounds.height / 2.0 + offsetY
+            offsetX: x + offsetX / scaleX,
+            offsetY: y + offsetY / scaleY,
+            size: sizeInPoints
         )
     }